fix(dataloader): add dataloader context to integration tests
- Move IDataLoaders interface from abstract types to graphql adapter
- Add DataLoaderService to GraphQL context in test setup
- Include necessary domain modules for repository access in tests
- Fix failing component and page resolver integration tests
