+++ libs/backend/abstract/types/src/server.interface.ts
@@ -1,8 +1,7 @@
+import type { IDataLoaders } from '@codelab/backend-infra-adapter-graphql'
 import type { Auth0IdToken } from '@codelab/shared-abstract-core'
 import type { IncomingMessage } from 'http'
 import type { NextApiRequest as OriginalNextApiRequest } from 'next'
-
-import type { IDataLoaders } from './dataloader.interface'
 
 export interface GraphQLRequestContext {
   req: NextApiRequest
