+++ libs/backend/infra/adapter/graphql/src/dataloader/dataloader.interface.ts
@@ -1,5 +1,16 @@
-// Re-export from backend-abstract-types to maintain backward compatibility
-export type {
-  IDataLoaderContext,
-  IDataLoaders,
-} from '@codelab/backend-abstract-types'
+import type {
+  ElementFragment,
+  TypeFragment,
+} from '@codelab/shared-infra-gqlgen'
+import type DataLoader from 'dataloader'
+import type { GraphQLResolveInfo } from 'graphql'
+
+export interface IDataLoaders {
+  elementDependantTypesLoader: DataLoader<string, Array<TypeFragment>>
+  pageElementsLoader: DataLoader<string, Array<ElementFragment>>
+}
+
+export interface IDataLoaderContext {
+  info?: GraphQLResolveInfo
+  loaders: IDataLoaders
+}
